requirements-dash.txt
dash>=2.14.0
dash-bootstrap-components>=1.5.0
google-cloud-bigquery>=3.20.0
pandas>=2.0.0
db-dtypes>=1.1.0
